五子棋升级版一代，先手战胜普通室友，自学习五代
