+++ experiment Pro/AI 五子棋/AI 五子棋.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31112.23
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AI 五子棋", "AI 五子棋.vcxproj", "{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Debug|x64.ActiveCfg = Debug|x64
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Debug|x64.Build.0 = Debug|x64
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Debug|x86.ActiveCfg = Debug|Win32
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Debug|x86.Build.0 = Debug|Win32
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Release|x64.ActiveCfg = Release|x64
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Release|x64.Build.0 = Release|x64
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Release|x86.ActiveCfg = Release|Win32
+		{5EEEB4EA-1563-4138-B1AB-B2BEC81D57AA}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {336AF65C-EBB5-450B-9CF3-637B7C16467C}
+	EndGlobalSection
+EndGlobal
