+++ experiment Pro/AI 五子棋/源.cpp
@@ -1,0 +1,9 @@
+#include<graphics.h>
+#include<math.h>
+#include<iostream>
+using namespace std;
+int main()
+{
+	cout << "hello world";
+	return 0;
+}
