--- experiment Pro/best_parameter.txt
@@ -1,2 +1,0 @@
-generation=0:
-0:0.000000 1:0.000000 2:0.000000 3:0.000000 4:0.000000 5:0.000000 6:0.000000
