+++ experiment Pro/head.cpp
@@ -1,0 +1,13 @@
+#include <bits/stdc++.h>
+#include <conio.h>
+#include <ctime>
+#include <windows.h>
+#ifndef consider_step
+#define consider_step 4 //考虑步数
+#define total_num 10    //棋手人数
+#define size 15         //棋盘大小
+#define generations 10  //进化代数
+#define mod (1 + total_num / 2) * total_num / 4
+using namespace std;
+int previousx, previousy;
+#endif
